Merge upstream changes

+++ tests/test_delete_protection.py
@@ -37,7 +37,8 @@ class TestDeleteProtection(TransactionTestCase):
             return self.obj
 
         def __exit__(self, exc_type, exc_val, exc_tb):
-            self.obj.delete()
+            if not exc_tb:
+                self.obj.delete()
 
     @property
     def scratchPrizeTimed(self):
@@ -106,7 +107,7 @@ class TestDeleteProtection(TransactionTestCase):
 
     @property
     def scratchBidRun(self):
-        return models.Bid.objects.get_or_create(name='Scratch Bid', speedrun=self.scratchRun)[0]
+        return models.Bid.objects.get_or_create(name='Scratch Bid', speedrun=self.scratchRun, istarget=True)[0]
 
     def testDeleteEvent(self):
         self.assertDeleteProtected(self.event, self.scratchBidEvent)
